fix. Order 내부의 더미 member 삭제, AuthMember로 교체 완료

+++ src/main/java/com/sparta/deliveryapp/domain/member/dto/AuthMember.java
@@ -6,11 +6,11 @@ import lombok.Getter;
 
 @Getter
 public class AuthMember {
-    private final Integer id;
+    private final Long id;
     private final String username;
     private final UserRole role;
 
-    public AuthMember(Integer id, String username, UserRole role){
+    public AuthMember(Long id, String username, UserRole role){
         this.id = id;
         this.username = username;
         this.role = role;
